fix: remodel example
