+++ src/config/sdk.py
@@ -1,0 +1,7 @@
+from lib.models import OutputFormat
+
+entries: int = 1
+profile: str
+format: OutputFormat
+debug: bool
+verbose: bool
