+++ src/lib/dummy_aws_credentials
