+++ src/main.py
@@ -1,23 +1,23 @@
 import sys
+from collections.abc import Sequence, Mapping
 from configparser import RawConfigParser
 from datetime import timedelta
 from enum import Enum
 from json import dumps
 from stat import S_IRUSR, S_IWUSR
-from typing import Iterable, Iterator, Dict, Optional, List
+from typing import Iterable, Iterator, Dict, Any, Optional, List
 from uuid import UUID
 from sys import version as python_version, stderr
 from time import perf_counter
+from lib.models import AlertStateSetable, OutputFormat, AWSAccount
+import config.sdk as skd_config
+
 import logging
-from utils import OutputFormat, output_iterable, dump_json 
-import options
-from account import Account
 
 import boto3
 from boto3_type_annotations.organizations import Client as Boto3OrganizationsClient
 import typer
-import click
-from click import Context
+from prettytable import PrettyTable
 from typer import Typer
 
 from zanshinsdk import Client, AlertState, AlertSeverity, Languages, AlertsOrderOpts, SortOpts, __version__ as sdk_version
@@ -26,32 +26,8 @@ from zanshinsdk.client import ScanTargetKind, ScanTargetSchedule, ScanTargetAWS,
 from zanshinsdk.alerts_history import FilePersistentAlertsIterator
 from zanshinsdk.following_alerts_history import FilePersistentFollowingAlertsIterator
 
-from awsorgrun import AWSOrgRunTarget, awsorgrun
-from version import __version__ as cli_version
-
-
-# Alert states that the user can set.
-class AlertStateSetable(str, Enum):
-    OPEN = "OPEN"
-    IN_PROGRESS = "IN_PROGRESS"
-    RISK_ACCEPTED = "RISK_ACCEPTED"
-    MITIGATING_CONTROL = "MITIGATING_CONTROL"
-    FALSE_POSITIVE = "FALSE_POSITIVE"
-
-
-class OrderedCommands(click.Group):
-    def list_commands(self, ctx: Context) -> Iterable[str]:
-        return self.commands.keys()
-
-class AWSAccount(dict):
-    """
-    Class representing a AWS Account as returned by boto3
-    """
-
-    def __init__(self, Id: str, Name: str, Arn: str, Email: str, Onboard: bool = False):
-        dict.__init__(self, Id=Id, Name=Name, Arn=Arn,
-                      Email=Email, Onboard=Onboard)
-
+from lib.awsorgrun import AWSOrgRunTarget, awsorgrun
+from lib.version import __version__ as cli_version
 
 ###################################################
 # Exchanger
@@ -65,9 +41,77 @@ sys.excepthook = zanshin_exchanger
 
 
 ###################################################
+# Utils
+###################################################
+
+def format_field(value: Any) -> str:
+    """
+    Function that formats a single field for output on a table or CSV output, in order to deal with nested arrays or
+    objects in the JSON outputs of the API
+    :param value: the value to format
+    :return: a string that is fit for console output
+    """
+    if isinstance(value, Sequence) and not isinstance(value, (str, bytes)):
+        if all(isinstance(x, (str, bytes, int, float)) for x in value):
+            return ", ".join([str(x) for x in value])
+        else:
+            return dumps(value)
+    elif isinstance(value, Mapping):
+        return dumps(value)
+    else:
+        return value
+
+
+def output_iterable(iterator: Iterator[Dict], empty: Any = None, _each_iteration_function: Any = None) -> None:
+    """
+    Function that iterates over a series of dicts representing JSON objects returned by API list operations, and which
+    outputs them using typer.echo in the specified format. Will use streaming processing for JSON, all others need to
+    load all responses in memory in a PrettyTable prior to output, which could be problematic for large number of
+    entries
+    :param _each_iteration_function:
+    :param empty:
+    :param iterator: the iterator containing the JSON objects
+    :return: None
+    """
+    global global_options
+
+    global_options['entries'] = 0
+    if global_options['format'] is OutputFormat.JSON:
+        for entry in iterator:
+            typer.echo(dumps(entry, indent=4))
+            global_options['entries'] += 1
+            if _each_iteration_function:
+                _each_iteration_function()
+    else:
+        table = PrettyTable()
+        for entry in iterator:
+            if not table.field_names:
+                table.field_names = sorted(entry.keys())
+            else:
+                for k in entry.keys():
+                    if k not in table.field_names:
+                        table.add_column(k, [empty] * global_options['entries'])
+            table.add_row([format_field(entry.get(fn, empty)) for fn in table.field_names])
+            global_options['entries'] += 1
+            if _each_iteration_function:
+                _each_iteration_function()
+        if global_options['format'] is OutputFormat.TABLE:
+            typer.echo(table.get_string())
+        elif global_options['format'] is OutputFormat.CSV:
+            typer.echo(table.get_csv_string())
+        elif global_options['format'] is OutputFormat.HTML:
+            typer.echo(table.get_html_string())
+        else:
+            raise NotImplementedError(f"unexpected format type {global_options['format']}")
+
+
+
+
+###################################################
 # Main App
 ###################################################
 
+global_options: dict = {'entries': 1}
 main_app: Typer = typer.Typer(name="zanshin", no_args_is_help=True)
 
 
@@ -92,7 +136,7 @@ def global_options_callback(ctx: typer.Context,
 
         def print_elapsed_time():
             typer.echo(
-                f"zanshin: {options.global_options['entries']} object(s) processed in "
+                f"zanshin: {skd_config.entries} object(s) processed in "
                 f"{timedelta(seconds=perf_counter() - start_time)}",
                 err=True)
 
@@ -108,12 +152,11 @@ def global_options_callback(ctx: typer.Context,
         handler.setFormatter(formatter)
         logger.addHandler(handler)
 
-    options.global_options['debug'] = debug
-    options.global_options['verbose'] = verbose
-    options.global_options['profile'] = profile
-    options.global_options['format'] = output_format
-
-
+    skd_config.debug = debug
+    skd_config.verbose = verbose
+    skd_config.profile = profile
+    skd_config.format = output_format
+    
 @main_app.command()
 def init():
     """
@@ -122,7 +165,7 @@ def init():
     cfg = RawConfigParser()
     cfg.read(CONFIG_FILE)
     typer.echo("This command will allow you to set up profiles in the configuration file.")
-    profile = typer.prompt("Please enter the profile name to use", default=options.global_options['profile'])
+    profile = typer.prompt("Please enter the profile name to use", default=global_options['profile'])
     if cfg.has_section(profile):
         typer.confirm("Profile already exists. Overwrite?", abort=True)
     else:
@@ -147,10 +190,10 @@ def version():
 ###################################################
 # Account App
 ###################################################
-account_app = Account().load_commands()
-main_app.add_typer(account_app, name="account",
-                   help="Operations on user the API key owner has direct access to")
 
+import bin.account as account
+main_app.add_typer(account.app, name="account",
+                   help="Operations on user the API key owner has direct access to")
 
 
 ###################################################
@@ -158,7 +201,7 @@ main_app.add_typer(account_app, name="account",
 ###################################################
 
 invites_app = typer.Typer()
-account_app.add_typer(invites_app, name="invites",
+account.app.add_typer(invites_app, name="invites",
                       help="Operations on invites from account the API key owner has direct access to")
 
 
@@ -167,7 +210,7 @@ def account_invite_list():
     """
     Iterates over the invites of current logged user.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_invites())
 
 
@@ -176,7 +219,7 @@ def account_invite_get(invite_id: UUID = typer.Argument(..., help="UUID of the i
     """
     Gets a specific invitation details, it only works if the invitation was made for the current logged user.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_invite(invite_id))
 
 
@@ -186,7 +229,7 @@ def account_invite_accept(invite_id: UUID = typer.Argument(..., help="UUID of th
     Accepts an invitation with the informed ID, it only works if the user accepting the invitation is the user that
     received the invitation.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_invite(invite_id))
 
 
@@ -195,7 +238,7 @@ def account_invite_accept(invite_id: UUID = typer.Argument(..., help="UUID of th
 ###################################################
 
 api_key_app = typer.Typer()
-account_app.add_typer(api_key_app, name="api_key",
+account.app.add_typer(api_key_app, name="api_key",
                       help="Operations on API keys from account the API key owner has direct access to")
 
 
@@ -204,7 +247,7 @@ def account_api_key_list():
     """
     Iterates over the API keys of current logged user.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_api_keys())
 
 
@@ -214,7 +257,7 @@ def account_api_key_create(name: str = typer.Argument(..., help="Name of the new
     Creates a new API key for the current logged user, API Keys can be used to interact with the zanshin api directly
     a behalf of that user.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.create_api_key(name))
 
 
@@ -223,7 +266,7 @@ def account_api_key_delete(api_key_id: UUID = typer.Argument(..., help="UUID of 
     """
     Deletes a given API key by its id, it will only work if the informed ID belongs to the current logged user.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_api_key(api_key_id))
 
 
@@ -241,7 +284,7 @@ def organization_list():
     """
     Lists the organizations this user has direct access to as a member.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organizations())
 
 
@@ -250,7 +293,7 @@ def organization_get(organization_id: UUID = typer.Argument(..., help="UUID of t
     """
     Gets an organization given its ID.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization(organization_id))
 
 
@@ -264,7 +307,7 @@ def organization_update(
     """
     Gets an organization given its ID.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.update_organization(organization_id, name, picture, email))
 
 
@@ -275,7 +318,7 @@ def organization_create(
     """
     Creates an organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.create_organization(name))
 
 @organization_app.command(name='delete')
@@ -283,7 +326,7 @@ def organization_delete(organization_id: UUID = typer.Argument(..., help="UUID o
     """
     Deletes an organization given its ID.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_organization(organization_id))
 
 
@@ -302,7 +345,7 @@ def organization_member_list(organization_id: UUID = typer.Argument(..., help="U
     """
     Lists the members of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_members(organization_id))
 
 
@@ -314,7 +357,7 @@ def organization_member_get(
     """
     Get organization member.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_member(organization_id, organization_member_id))
 
 
@@ -329,7 +372,7 @@ def organization_member_update(
     """
     Update organization member.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.update_organization_member(organization_id, organization_member_id, role))
 
 
@@ -341,7 +384,7 @@ def organization_member_delete(
     """
     Delete organization member.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_organization_member(organization_id, organization_member_id))
 
 
@@ -360,7 +403,7 @@ def organization_member_invite_list(organization_id: UUID = typer.Argument(..., 
     """
     Lists the member invites of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_members_invites(organization_id))
 
 
@@ -375,7 +418,7 @@ def organization_member_invite_create(
     """
     Create organization member invite.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.create_organization_members_invite(organization_id, organization_member_invite_email,
                                                         organization_member_invite_role))
 
@@ -388,7 +431,7 @@ def organization_member_invite_get(
     """
     Get organization member invite.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_member(organization_id, organization_member_invite_email))
 
 
@@ -400,7 +443,7 @@ def organization_member_invite_delete(
     """
     Delete organization member invite.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_organization_member_invite(organization_id, organization_member_invite_email))
 
 
@@ -412,7 +455,7 @@ def organization_member_invite_resend(
     """
     Resend organization member invitation.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.resend_organization_member_invite(organization_id, organization_member_invite_email))
 
 
@@ -430,7 +473,7 @@ def organization_follower_list(organization_id: UUID = typer.Argument(..., help=
     """
     Lists the followers of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_followers(organization_id))
 
 
@@ -442,7 +485,7 @@ def organization_follower_stop(
     """
     Stops one organization follower of another.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.stop_organization_follower(organization_id, organization_follower_id))
 
 
@@ -461,7 +504,7 @@ def organization_follower_request_list(organization_id: UUID = typer.Argument(..
     """
     Lists the follower requests of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_followers(organization_id))
 
 
@@ -473,7 +516,7 @@ def organization_follower_request_create(
     """
     Create organization follower request.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.create_organization_follower_request(organization_id, token))
 
 
@@ -485,7 +528,7 @@ def organization_follower_request_get(
     """
     Get organization follower request.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_follower_request(organization_id, token))
 
 
@@ -497,7 +540,7 @@ def organization_follower_request_delete(
     """
     Delete organization follower request.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_organization_follower_request(organization_id, token))
 
 
@@ -515,7 +558,7 @@ def organization_following_list(organization_id: UUID = typer.Argument(..., help
     """
     Lists the following of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_following(organization_id))
 
 
@@ -527,7 +570,7 @@ def organization_following_stop(
     """
     Stops one organization following of another.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.stop_organization_following(organization_id, organization_following_id))
 
 
@@ -546,7 +589,7 @@ def organization_following_request_list(organization_id: UUID = typer.Argument(.
     """
     Lists the following requests of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_following_requests(organization_id))
 
 
@@ -558,7 +601,7 @@ def organization_following_request_get(
     """
     Returns a request received by an organization to follow another.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_following_request(organization_id, following_id))
 
 
@@ -570,7 +613,7 @@ def organization_following_request_accept(
     """
     Accepts a request to follow another organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.accept_organization_following_request(organization_id, following_id))
 
 
@@ -582,7 +625,7 @@ def organization_following_request_decline(
     """
     Declines a request to follow another organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.decline_organization_following_request(organization_id, following_id))
 
 
@@ -600,7 +643,7 @@ def organization_scan_target_list(organization_id: UUID = typer.Argument(..., he
     """
     Lists the scan targets of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_scan_targets(organization_id))
 
 
@@ -615,7 +658,7 @@ def organization_scan_target_create(
     """
     Create a new scan target in organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     if kind == ScanTargetKind.AWS:
         credential = ScanTargetAWS(credential)
     elif kind == ScanTargetKind.GCP:
@@ -637,7 +680,7 @@ def organization_scan_target_get(
     """
     Get scan target of organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_scan_target(organization_id, scan_target_id))
 
 
@@ -651,7 +694,7 @@ def organization_scan_target_update(
     """
     Update scan target of organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.update_organization_scan_target(organization_id, scan_target_id, name, schedule))
 
 
@@ -663,7 +706,7 @@ def organization_scan_target_delete(
     """
     Delete scan target of organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_organization_scan_target(organization_id, scan_target_id))
 
 
@@ -675,7 +718,7 @@ def organization_scan_target_check(
     """
     Check scan target.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.check_organization_scan_target(organization_id, scan_target_id))
 
 @organization_scan_target_app.command(name='oauth-link')
@@ -686,7 +729,7 @@ def organization_scan_target_oauth_link(
     """
     Retrieve a link to allow the user to authorize zanshin to read info from their gworkspace environment.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_gworkspace_oauth_link(organization_id, scan_target_id))
 
 @organization_scan_target_app.command(name='onboard_aws')
@@ -702,7 +745,7 @@ def onboard_organization_aws_scan_target(
     Create a new scan target in organization and perform onboard. Requires boto3 and correct AWS IAM Privileges.
     Checkout the required AWS IAM privileges here https://github.com/tenchi-security/zanshin-sdk-python/blob/main/zanshinsdk/docs/README.md
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     credential = ScanTargetAWS(credential)
     kind = ScanTargetKind.AWS
 
@@ -742,7 +785,7 @@ def onboard_organization_aws_organization_scan_target(
     Checkout the required AWS IAM privileges at
     https://github.com/tenchi-security/zanshin-cli/blob/main/zanshincli/docs/README.md
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     if boto3_profile:
         boto3_session = boto3.Session(profile_name=boto3_profile)
     else:
@@ -818,7 +861,7 @@ def onboard_organization_aws_organization_scan_target(
 
 def _sdk_onboard_scan_target(target, aws_account_id, aws_account_name, boto3_session, region, organization_id,
                              schedule):
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     account_credential = ScanTargetAWS(aws_account_id)
     client.onboard_scan_target(boto3_session=boto3_session, region=region, kind=ScanTargetKind.AWS,
                                name=aws_account_name,
@@ -887,7 +930,7 @@ def organization_scan_target_scan_start(
     """
     Starts a scan on the specified scan target.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.start_organization_scan_target_scan(organization_id, scan_target_id, force))
 
 
@@ -899,7 +942,7 @@ def organization_scan_target_scan_stop(
     """
     Stop a scan on the specified scan target.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.stop_organization_scan_target_scan(organization_id, scan_target_id))
 
 
@@ -911,7 +954,7 @@ def organization_scan_target_scan_list(
     """
     Lists the scan target scans of organization this user has direct access to.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_scan_target_scans(organization_id, scan_target_id))
 
 
@@ -924,7 +967,7 @@ def organization_scan_target_scan_get(
     """
     Get scan of scan target.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_scan_target_scan(organization_id, scan_target_id, scan_id))
 
 ###################################################
@@ -943,7 +986,7 @@ def scan_target_groups_script(
     """
     Gets download URL of the scan target group.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_scan_target_group_script(organization_id, scan_target_group_id))
 
 @scan_target_group_app.command(name='compartments')
@@ -954,7 +997,7 @@ def scan_target_groups_compartments(
     """
     Iterates over the compartments of a scan target group.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_scan_target_group_compartments(organization_id, scan_target_group_id))
 
 @scan_target_group_app.command(name='insert')
@@ -970,7 +1013,7 @@ def scan_target_groups_insert(
     Inserts an already created scan target group.
     """
     credential = ScanTargetGroupCredentialListORACLE(region, tenancy_id, user_id, key_fingerprint)
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.insert_scan_target_group_credential(organization_id, scan_target_group_id,credential))
 
 @scan_target_group_app.command(name='create-by-compartments')
@@ -983,7 +1026,7 @@ def scan_target_groups_create_by_compartments(
     """
     Creates Scan Targets from previous listed compartments inside the scan target group.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.create_scan_target_by_compartments(organization_id, scan_target_group_id,name, ocid))
 
 @scan_target_group_app.command(name='scan-targets')
@@ -994,7 +1037,7 @@ def scan_target_groups_scan_targets(
     """
     Gets all scan targets from a specific scan target group.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_scan_targets_from_group(organization_id, scan_target_group_id))
                    
 @scan_target_group_app.command(name='get')
@@ -1005,7 +1048,7 @@ def scan_target_groups_get(
     """
     Gets details of the scan target group given its ID.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.get_organization_scan_target_group(organization_id, scan_target_group_id))
 
 @scan_target_group_app.command(name='delete')
@@ -1016,7 +1059,7 @@ def scan_target_groups_delete(
     """
     Deletes the scan target group of the organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.delete_organization_scan_target_group(organization_id, scan_target_group_id))
 
 @scan_target_group_app.command(name='list')
@@ -1024,7 +1067,7 @@ def scan_target_groups_list(organization_id: UUID = typer.Argument(..., help="UU
     """
     Lists the scan target groups of the user's organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(client.iter_organization_scan_target_groups(organization_id))
 
 
@@ -1037,7 +1080,7 @@ def scan_target_groups_update(
     """
     Updates a scan target group.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.update_scan_target_group(organization_id, scan_target_group_id,name))
 
 
@@ -1050,7 +1093,7 @@ def scan_target_groups_create(
     """
     Creates a scan target group for the organization.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     dump_json(client.create_scan_target_group(organization_id, kind, name))
 
 
@@ -1089,7 +1132,7 @@ def alert_list(organization_id: UUID = typer.Argument(..., help="UUID of the org
     """
     List alerts from a given organization, with optional filters by scan target, state or severity.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(
         client.iter_alerts(organization_id=organization_id, scan_target_ids=scan_target_id, states=states,
                            severities=severity, language=language,
@@ -1125,7 +1168,7 @@ def alert_following_list(organization_id: UUID = typer.Argument(..., help="UUID 
     """
     List following alerts from a given organization, with optional filters by following ids, state or severity.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(
         client.iter_following_alerts(organization_id=organization_id, following_ids=following_ids, states=states,
                                      created_at_start=created_at_start, created_at_end=created_at_end,
@@ -1145,7 +1188,7 @@ def alert_history_list(organization_id: UUID = typer.Argument(..., help="UUID of
     """
     List alerts from a given organization, with optional filters by scan target, state or severity.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
 
     if persist:
         iter_alerts = FilePersistentAlertsIterator(filename='zanshin', client=client, organization_id=organization_id,
@@ -1172,7 +1215,7 @@ def alert_history_following_list(organization_id: UUID = typer.Argument(..., hel
     """
     List alerts from a given organization, with optional filters by scan target, state or severity.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
 
     if persist:
         iter_alerts = FilePersistentFollowingAlertsIterator(filename='zanshin', client=client,
@@ -1205,7 +1248,7 @@ def grouped_alert_list(organization_id: UUID = typer.Argument(..., help="UUID of
     """
     List grouped alerts from a given organization, with optional filters by scan target, state or severity.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(
         client.iter_grouped_alerts(organization_id=organization_id, scan_target_ids=scan_target_id, states=state,
                                    severities=severity))
@@ -1226,7 +1269,7 @@ def grouped_alert_following_list(organization_id: UUID = typer.Argument(..., hel
     """
     List grouped following alerts from a given organization, with optional filters by scan target, state or severity.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     output_iterable(
         client.iter_grouped_following_alerts(organization_id=organization_id, following_ids=following_ids, states=state,
                                              severities=severity))
@@ -1240,13 +1283,13 @@ def alert_get(alert_id: UUID = typer.Argument(..., help="UUID of the alert to lo
     Returns details about a specified alert
     """
     if list_history:
-        client = Client(profile=options.global_options['profile'])
+        client = Client(profile=global_options['profile'])
         output_iterable(client.iter_alert_history(alert_id))
     elif list_comments:
-        client = Client(profile=options.global_options['profile'])
+        client = Client(profile=global_options['profile'])
         output_iterable(client.iter_alert_comments(alert_id))
     else:
-        client = Client(profile=options.global_options['profile'])
+        client = Client(profile=global_options['profile'])
         typer.echo(dumps(client.get_alert(alert_id), indent=4))
 
 
@@ -1262,7 +1305,7 @@ def alert_update(organization_id: UUID = typer.Argument(..., help="UUID of the o
     Updates the alert.
     """
 
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
     typer.echo(client.update_alert(organization_id, scan_target_id, alert_id, state, labels, comment))
 
 
@@ -1283,7 +1326,7 @@ def summary_alert(organization_id: UUID = typer.Argument(..., help="UUID of the 
     """
     Gets a summary of the current state of alerts for an organization, both in total and broken down by scan target.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
 
     dump_json(client.get_alert_summaries(organization_id, scan_target_id))
 
@@ -1297,7 +1340,7 @@ def summary_alert_following(organization_id: UUID = typer.Argument(..., help="UU
     """
     Gets a summary of the current state of alerts for followed organizations.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
 
     dump_json(client.get_following_alert_summaries(organization_id, following_ids))
 
@@ -1312,7 +1355,7 @@ def summary_scan(organization_id: UUID = typer.Argument(..., help="UUID of the o
     """
     Returns summaries of scan results over a period of time, summarizing number of alerts that changed states.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
 
     dump_json(client.get_scan_summaries(organization_id, scan_target_ids, days))
 
@@ -1330,7 +1373,7 @@ def summary_scan_following(organization_id: UUID = typer.Argument(..., help="UUI
     """
     Returns summaries of scan results over a period of time, summarizing number of alerts that changed states.
     """
-    client = Client(profile=options.global_options['profile'])
+    client = Client(profile=global_options['profile'])
 
     dump_json(client.get_scan_summaries(organization_id, following_ids, days))
