--- zanshincli/src/options.py
@@ -1,1 +1,0 @@
-global_options: dict = {'entries': 1}
